// src/components/supplier/types/edit-supplier.ts
export interface EditSupplierPayload {
    nombre: string
    correo: string
    telefono: string
    tipoDocumento: string
    departamento: string
    nombreDepartamento: string
    nombreComercial: string
    nit: string
    nrc: string
    numDocumento: string
    esContribuyente: number
    municipio: string
    nombreMunicipio: string
    codCuenta: string
    codActividad: string
    descActividad: string
    complemento: string
}